src/test/java/com/github/brunomndantas/jscrapper/support/parser/single/attribute/reference/SingleLongAttributeParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.single.attribute.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class SingleLongAttributeParserTest{

    @Test
    public void constructorTest() {
        String attribute = "";
        SingleLongAttributeParser parser = new SingleLongAttributeParser(attribute);
        assertEquals(null, parser.getDefaultValue());
        assertSame(attribute, parser.getAttribute());
    }

    @Test
    public void parseElement() throws Exception {
        SingleLongAttributeParser parser = new SingleLongAttributeParser("");

        assertEquals(2L, parser.parseElement(new DummyElement(){
            @Override
            public String getAttribute(String a) {
                return "2";
            }
        }));
    }

}
